+++ README.md
@@ -1,19 +1,15 @@
-# Kestra Plugin Template
+# Kestra Storage for Amazon S3
 
 <p align="center">
   <img width="460" src="https://kestra.io/logo.svg"  alt="Kestra workflow orchestrator" />
 </p>
 
-> A template for creating plugin for Kestra
+> A storage for Kestra using S3
 
-This repository is meant to serve as a general template for how to set up new plugin for 
-[Kestra](https://github.com/kestra-io/kestra). In general, setting up a new plugin should 
-take only a few minutes; use this repository as a way of finding example files to ensure 
-that you've set up the plugin correctly.
+![Kestra orchestrator](https://kestra.io/ui.gif)
 
 ## Documentation
 * The official Kestra documentation can be found under: [kestra.io](https://kestra.io)
-* The official documentation for developing a plugin can be found [here](https://kestra.io/docs/plugin-developer-guide/)
 
 ## License
 Apache 2.0 © [Kestra Technologies](https://kestra.io)
