+++ src/main/java/io/kestra/storage/s3/S3Config.java
@@ -1,16 +1,16 @@
-package io.kestra.storage.s3.config;
+package io.kestra.storage.s3;
 
 import io.micronaut.context.annotation.ConfigurationProperties;
 import jakarta.inject.Singleton;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
 @Singleton
 @Getter
+@Builder
 @ConfigurationProperties("kestra.storage.s3")
-public class S3Properties {
-
-	@Setter
+public class S3Config {
 	private String bucket;
 
 	private String region;
