+++ src/main/java/io/kestra/storage/s3/S3StorageEnabled.java
@@ -1,4 +1,4 @@
-package io.kestra.storage.s3.config;
+package io.kestra.storage.s3;
 
 import io.micronaut.context.annotation.Requires;
